Extract table
