+++ frontend/src/features/admins/AdminReceptionistsPage.tsx
@@ -1,9 +1,7 @@
 ﻿import { useForm } from "@tanstack/react-form";
 import { useQueryClient } from "@tanstack/react-query";
 import { Link } from "@tanstack/react-router";
-import { type ColumnDef, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
 import { motion } from "framer-motion";
-import { ChevronLeft, ChevronRight, Eye, Pencil, Plus, Search, UserX } from "lucide-react";
 import { useEffect, useState } from "react";
 import { toast } from "sonner";
 import { z } from "zod";
@@ -43,14 +41,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Separator } from "@/components/ui/separator";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
+import { ReceptionistTable } from "@/features/admins/manageReceptionists/ReceptionistTable";
 
 const ACCENT = "#0d9488";
 const PAGE_SIZE = 10;
@@ -1107,267 +1098,6 @@ function DeactivateReceptionistDialog({
   );
 }
 
-// Column definitions for the receptionists data table
-const columns: ColumnDef<AdminReceptionistDto>[] = [
-  {
-    accessorKey: "firstName",
-    header: "Name",
-    cell: ({ row }) => (
-      <div className="flex flex-col gap-0.5">
-        <span className="font-medium">
-          {row.original.firstName} {row.original.lastName}
-        </span>
-        <span className="font-mono text-xs text-muted-foreground">@{row.original.username}</span>
-      </div>
-    ),
-  },
-  {
-    accessorKey: "email",
-    header: "Email",
-    cell: ({ row }) => (
-      <span className="text-muted-foreground text-xs">{row.getValue("email")}</span>
-    ),
-  },
-  {
-    accessorKey: "phoneNumber",
-    header: "Phone",
-    cell: ({ row }) => (
-      <span className="font-mono text-xs">{row.getValue("phoneNumber") || "—"}</span>
-    ),
-  },
-  {
-    accessorKey: "createdAt",
-    header: "Joined",
-    cell: ({ row }) => (
-      <span className="text-xs text-muted-foreground">{formatDate(row.getValue("createdAt"))}</span>
-    ),
-  },
-  {
-    id: "actions",
-    header: "Actions",
-    cell: ({ row, table }) => (
-      <div className="flex items-center gap-1">
-        <Button
-          variant="ghost"
-          size="sm"
-          className="h-8 w-8 p-0 text-muted-foreground hover:text-foreground"
-          title="View details"
-          onClick={() =>
-            (
-              table.options.meta as {
-                onView: (r: AdminReceptionistDto) => void;
-                onEdit: (r: AdminReceptionistDto) => void;
-              }
-            ).onView(row.original)
-          }
-        >
-          <Eye className="size-3.5" />
-        </Button>
-        <Button
-          variant="ghost"
-          size="sm"
-          className="h-8 w-8 p-0 text-muted-foreground hover:text-foreground"
-          title="Edit receptionist"
-          onClick={() =>
-            (
-              table.options.meta as {
-                onView: (r: AdminReceptionistDto) => void;
-                onEdit: (r: AdminReceptionistDto) => void;
-              }
-            ).onEdit(row.original)
-          }
-        >
-          <Pencil className="size-3.5" />
-        </Button>
-        <Button
-          variant="ghost"
-          size="sm"
-          className="h-8 w-8 p-0 text-muted-foreground hover:text-destructive"
-          title="Remove receptionist"
-          onClick={() =>
-            (
-              table.options.meta as {
-                onView: (r: AdminReceptionistDto) => void;
-                onEdit: (r: AdminReceptionistDto) => void;
-                onDeactivate: (r: AdminReceptionistDto) => void;
-              }
-            ).onDeactivate(row.original)
-          }
-        >
-          <UserX className="size-3.5" />
-        </Button>
-      </div>
-    ),
-  },
-];
-
-interface DataTableProps {
-  data: AdminReceptionistDto[];
-  page: number;
-  totalPages: number;
-  hasNextPage?: boolean;
-  hasPreviousPage?: boolean;
-  onPageChange: (page: number) => void;
-  search: string;
-  onSearchChange: (value: string) => void;
-  onView: (receptionist: AdminReceptionistDto) => void;
-  onEdit: (receptionist: AdminReceptionistDto) => void;
-  onDeactivate: (receptionist: AdminReceptionistDto) => void;
-  onAddNew: () => void;
-}
-
-// Paginated and searchable data table for the receptionists list
-function DataTable({
-  data,
-  page,
-  totalPages,
-  hasNextPage,
-  hasPreviousPage,
-  onPageChange,
-  search,
-  onSearchChange,
-  onView,
-  onEdit,
-  onDeactivate,
-  onAddNew,
-}: DataTableProps) {
-  const table = useReactTable({
-    data,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-    meta: { onView, onEdit, onDeactivate },
-  });
-
-  const rows = table.getRowModel().rows;
-
-  return (
-    <div className="space-y-4">
-      <div className="flex items-center justify-between gap-4">
-        <div className="relative w-72">
-          <Search className="pointer-events-none absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-muted-foreground" />
-          <Input
-            placeholder="Search by name, email, department or ID…"
-            value={search}
-            onChange={(e) => onSearchChange(e.target.value)}
-            className="h-9 pl-9 text-sm"
-          />
-        </div>
-        <Button className="h-9 gap-1.5 bg-black text-white hover:bg-black/85" onClick={onAddNew}>
-          <Plus className="size-4" />
-          New Receptionist
-        </Button>
-      </div>
-
-      <div className="overflow-hidden rounded-lg border border-border">
-        <Table>
-          <TableHeader>
-            {table.getHeaderGroups().map((headerGroup) => (
-              <TableRow
-                key={headerGroup.id}
-                className="border-b border-foreground/20 bg-foreground hover:bg-foreground"
-              >
-                {headerGroup.headers.map((header) => (
-                  <TableHead
-                    key={header.id}
-                    className="px-5 py-3.5 text-[11px] font-semibold uppercase tracking-[0.15em] text-background/70"
-                  >
-                    {header.isPlaceholder
-                      ? null
-                      : flexRender(header.column.columnDef.header, header.getContext())}
-                  </TableHead>
-                ))}
-              </TableRow>
-            ))}
-          </TableHeader>
-          <TableBody>
-            {rows.length ? (
-              rows.map((row, i) => (
-                <TableRow
-                  key={row.id}
-                  className="border-b border-border transition-colors duration-100 hover:bg-muted/50"
-                  style={{ animationDelay: `${i * 30}ms` }}
-                >
-                  {row.getVisibleCells().map((cell) => (
-                    <TableCell key={cell.id} className="px-5 py-3.5 text-sm">
-                      {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                    </TableCell>
-                  ))}
-                </TableRow>
-              ))
-            ) : (
-              <TableRow>
-                <TableCell colSpan={columns.length} className="h-32 text-center">
-                  <p className="text-sm text-muted-foreground">No receptionists found.</p>
-                  <p className="mt-1 text-xs text-muted-foreground/60">
-                    Try adjusting your search.
-                  </p>
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-      </div>
-
-      {totalPages > 1 && (
-        <div className="flex items-center justify-between pt-1">
-          <p className="text-xs text-muted-foreground">
-            Page <span className="font-medium text-foreground">{page}</span> of{" "}
-            <span className="font-medium text-foreground">{totalPages}</span>
-          </p>
-          <div className="flex items-center gap-1.5">
-            <Button
-              variant="outline"
-              size="sm"
-              className="h-8 w-8 p-0"
-              disabled={!hasPreviousPage}
-              onClick={() => onPageChange(page - 1)}
-            >
-              <ChevronLeft className="size-4" />
-            </Button>
-
-            {Array.from({ length: totalPages }, (_, i) => i + 1)
-              .filter((p) => p === 1 || p === totalPages || (p >= page - 1 && p <= page + 1))
-              .reduce<(number | string)[]>((acc, p, idx, arr) => {
-                if (idx > 0 && p - (arr[idx - 1] as number) > 1)
-                  acc.push(`ellipsis-after-${arr[idx - 1]}`);
-                acc.push(p);
-                return acc;
-              }, [])
-              .map((item) =>
-                typeof item === "string" ? (
-                  <span key={item} className="px-1 text-xs text-muted-foreground">
-                    …
-                  </span>
-                ) : (
-                  <Button
-                    key={item}
-                    variant={item === page ? "default" : "outline"}
-                    size="sm"
-                    className="h-8 w-8 p-0 text-xs"
-                    style={item === page ? { background: ACCENT } : undefined}
-                    onClick={() => onPageChange(item)}
-                  >
-                    {item}
-                  </Button>
-                ),
-              )}
-
-            <Button
-              variant="outline"
-              size="sm"
-              className="h-8 w-8 p-0"
-              disabled={!hasNextPage}
-              onClick={() => onPageChange(page + 1)}
-            >
-              <ChevronRight className="size-4" />
-            </Button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
 // Admin page displaying a paginated, searchable list of all receptionists
 export function AdminReceptionistsPage() {
   const [page, setPage] = useState(1);
@@ -1463,7 +1193,7 @@ export function AdminReceptionistsPage() {
                 <p className="text-sm text-destructive">Failed to load receptionists.</p>
               </div>
             ) : (
-              <DataTable
+              <ReceptionistTable
                 data={receptionists}
                 page={page}
                 totalPages={totalPages}
